requirements.txt
alembic==1.0.2
arrow==0.15.4
astroid==2.2.0
autopep8==1.4.3
beautifulsoup4==4.6.0
blinker==1.4
bs4==0.0.1
cachetools==3.1.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.4.1
cycler==0.10.0
dominate==2.3.5
ecdsa==0.13
elasticsearch==6.3.1
feedparser==5.2.1
Flask==1.0.2
Flask-API==1.1
Flask-Avatars==0.2.1
Flask-Bootstrap==3.3.7.1
Flask-Email==1.4.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.14.0
idna==2.6
isort==4.3.9
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
numpy==1.14.2
pandas==0.22.0
Pillow==5.3.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycountry==18.2.23
pygal==2.4.0
pygal-maps-world==1.0.2
pygame==1.9.4
Pygments==2.3.1
PyJWT==1.6.4
pylint==2.3.0
pyparsing==2.2.0
pyping==0.0.5
python-dateutil==2.7.2
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.4
requests==2.18.4
rsa==4.0
scikit-learn==0.19.1
selenium==3.141.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.14
typed-ast==1.2.0
uritemplate==3.0.0
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.11.1
WTForms==2.2.1
yandex.translate==0.3.5
